Rename PlaygroundResultFeatureStrategyList to LegacyPlaygroundResultFeatureStrategyList
